+++ src/transport.rs
@@ -17,7 +17,7 @@ use crate::{
     encoding::{message::Message, Marshallable},
     peer::PeerNode,
     transport::encoding::{
-        Configurable, Decoder, Encoder, RaptorQDecoder, RaptorQEncoder,
+        Configurable, Decoder, Encoder, TransportDecoder, TransportEncoder,
     },
     MAX_DATAGRAM_SIZE,
 };
@@ -34,22 +34,27 @@ impl WireNetwork {
         inbound_channel_tx: Sender<MessageBeanIn>,
         listen_address: String,
         outbound_channel_rx: Receiver<MessageBeanOut>,
+        conf: HashMap<String, String>,
     ) {
         let listen_address = listen_address
             .parse()
             .expect("Unable to parse public_ip address");
-        let a = WireNetwork::listen_out(outbound_channel_rx);
-        let b =
-            WireNetwork::listen_in(listen_address, inbound_channel_tx.clone());
+        let a = WireNetwork::listen_out(outbound_channel_rx, &conf);
+        let b = WireNetwork::listen_in(
+            listen_address,
+            inbound_channel_tx.clone(),
+            &conf,
+        );
         let _ = tokio::join!(a, b);
     }
 
     async fn listen_in(
         listen_address: SocketAddr,
         inbound_channel_tx: Sender<MessageBeanIn>,
+        conf: &HashMap<String, String>,
     ) -> io::Result<()> {
         debug!("WireNetwork::listen_in started");
-        let mut decoder = RaptorQDecoder::configure(HashMap::new());
+        let mut decoder = TransportDecoder::configure(conf);
         let socket = UdpSocket::bind(listen_address)
             .await
             .expect("Unable to bind address");
@@ -90,9 +95,10 @@ impl WireNetwork {
 
     async fn listen_out(
         mut outbound_channel_rx: Receiver<MessageBeanOut>,
+        conf: &HashMap<String, String>,
     ) -> io::Result<()> {
         debug!("WireNetwork::listen_out started");
-        let encoder = RaptorQEncoder::configure(HashMap::new());
+        let encoder = TransportEncoder::configure(conf);
         loop {
             if let Some((message, to)) = outbound_channel_rx.recv().await {
                 trace!("< Message to send to ({:?}) - {:?} ", to, message);
@@ -124,4 +130,10 @@ impl WireNetwork {
         socket.send(data).await?;
         Ok(())
     }
+}
+
+pub fn default_configuration() -> HashMap<String, String> {
+    let mut conf = TransportEncoder::default_configuration();
+    conf.extend(TransportDecoder::default_configuration());
+    conf
 }